### 2025-07-24 8df3918
address paper pre-review comments
