+++ paper/paper.bib
@@ -4,6 +4,7 @@
   booktitle={International conference on Tools and Algorithms for the Construction and Analysis of Systems},
   pages={337--340},
   year={2008},
+  doi={10.1007/978-3-540-78800-3_24},
   organization={Springer}
 }
 
@@ -15,16 +16,8 @@
   number={2},
   pages={e2019MS001916},
   year={2020},
+  doi={10.1029/2019MS001916},
   publisher={Wiley Online Library}
-}
-
-@misc{ipywidgets,
-  author = {Jupyter},
-  title = {ipywidgets: Interactive HTML Widgets for Jupyter Notebooks},
-  year = {2015},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/jupyter-widgets/ipywidgets}},
 }
 
 @article{wu2021coupled,
@@ -35,6 +28,7 @@
   number={4},
   pages={e2020MS002418},
   year={2021},
+  doi={10.1029/2020MS002418},
   publisher={Wiley Online Library}
 }
 
@@ -45,6 +39,7 @@
   volume={99},
   number={3},
   pages={15--16},
+  doi={10.1029/2017EO079417},
   year={2017}
 }
 
@@ -56,6 +51,7 @@
   number={2},
   pages={250--280},
   year={2019},
+  doi={10.1029/2018RG000607},
   publisher={Wiley Online Library}
 }
 
@@ -67,13 +63,6 @@
   number={9},
   pages={69--77},
   year={2011},
+  doi={10.1145/1995376.1995394},
   publisher={ACM New York, NY, USA}
-}
-
-@book{biere2009handbook,
-  title={Handbook of satisfiability},
-  author={Biere, Armin and Heule, Marijn and van Maaren, Hans},
-  volume={185},
-  year={2009},
-  publisher={IOS press}
 }
