+++ paper/paper.md
@@ -1,12 +1,9 @@
 ---
 title: 'visualCaseGen: An SMT-based Experiment Configurator for Community Earth System Model'
 tags:
-  - Python
-  - cesm
-  - climate modeling
+  - climate model
+  - configuration
   - constraint solver
-  - product configuration
-  - SMT
 authors:
   - name: Alper Altuntas
     orcid: 0000-0003-1708-9518
@@ -41,204 +38,134 @@ bibliography: paper.bib
 
 ---
 
-# Introduction
+# Summary
 
-visualCaseGen is a graphical user interface (GUI) designed to streamline the
-creation and configuration of Community Earth System Model (CESM) experiments.
-CESM is a state-of-the-art climate model that allows researchers
-to simulate the Earth system across a wide range of spatial and temporal scales
-and at varying levels of complexity, depending on scientific objectives [@danabasoglu2020community].
-However, configuring non-standard or idealized CESM experiments is often a technically
-demanding and time-consuming process, requiring detailed knowledge of component
-compatibility, grid definitions, parameterization schemes, and model hierarchies.
-visualCaseGen simplifies this process by guiding users through each stage of
-experiment setup and automating key configuration steps via an intuitive, 
-interactive interface.
+visualCaseGen is a graphical user interface (GUI) that streamlines the setup of
+Community Earth System Model (CESM) experiments. While CESM supports simulations
+across a wide range of spatial and temporal scales and complexities [@danabasoglu2020community],
+configuring non-standard experiments can be technically demanding and time-consuming.
 
-To ensure consistency and compatibility across various model settings, visualCaseGen
-incorporates a constraint solver based on satisfiability modulo theories (SMT)
-[@de2011satisfiability]. This solver systematically analyzes dependencies between
-configuration parameters, detects conflicts, and provides detailed explanations of
-incompatibilities, allowing users to make informed adjustments. The SMT-based
-approach enables dynamic, real-time validation of model settings, significantly
-reducing setup errors and ensuring that only scientifically viable configurations
-are selected.  Complementing the solver, the visualCaseGen backend implements a
-*Stage Mechanism* that divides the workflow into structured, sequential steps. 
-This mechanism works in tandem with the solver to manage variable precedence,
-guide user input, and enforce constraints efficiently.
+To simplify this process, visualCaseGen offers a Jupyter-based interface 
+that guides users through each configuration stage and automates key tasks. 
+It features a constraint solver based on satisfiability modulo theories 
+(SMT) [@de2011satisfiability], which validates settings in real time by analyzing
+dependencies, detecting conflicts, and explaining incompatibilities. A complementary
+*Stage Mechanism* structures the workflow to manage variable precedence and to enforce
+constraints in a targeted and efficient manner.
 
-On the frontend, visualCaseGen is implemented as a Jupyter-based GUI [@ipywidgets], offering
-an intuitive, step-by-step interface for browsing standard CESM configurations,
-defining custom experiment setups, and modifying grid and component settings.
-Designed with a wizard-like interface, visualCaseGen walks users through each
-stage of the CESM configuration process, ensuring that all necessary settings
-are selected in a logical sequence while dynamically updating available options
-based on user choices. Additionally, the tool features utilities for creating
-and editing model input files, such as ocean grid, bathymetry, and land
-surface properties, further simplifying model customization and
-eliminating the need for manual file modifications.
-
-By automating and simplifying CESM configuration, visualCaseGen makes the model
-more accessible and custumizable, particularly for researchers exploring
-hierarchical modeling [@maher2019model], idealized experiments
-[@polvani2017less], or custom coupled simulations. As such, the tool allows
[…]
