### 2025-07-25 #8552 @alperaltuntas
@kthyng

- We have reduced the length of the paper in line with the paper format information.
- We don’t plan to release on PyPI, as conda- and GitHub-based installation better suit our use case. Installation details are available at: https://esmci.github.io/visualCaseGen/installation.html
