### 2025-07-29 #8552 @alperaltuntas
@kthyng

To ensure compatibility with specific CESM versions, visualCaseGen is bundled with CESM and distributed alongside it, rather than as a standalone package. Also, users often maintain multiple CESM versions on their systems, each used for different experiments or configurations. In such workflows, a single global installation of visualCaseGen (via PyPI or conda-forge) would not work across different CESM versions. Bundling it directly with CESM ensures version consistency and avoids potential conflicts or mismatches across environments.
