### 2025-08-22 #8552 @mengqi-z
👋 @andersy005 @zarzycki - Would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
