### 2025-09-16 #8552 @mengqi-z
👋 @dabail10 @sophmaca @duvivier - Would any of you be interested and available to review this submission for JOSS? Given your background, I believe you'd be a great fit to review this paper. JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
