### 2025-09-17 #8552 @dabail10
Alper is a software engineer here at NCAR and I actually work closely with him, so that rules out the three of us.
