### 2025-09-17 #8552 @mengqi-z
👋 @koldunovn @aaronspring - Would any of you be willing to review this submission for JOSS? JOSS carries out a checklist-driven review process here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
