### 2025-09-25 #8552 @andersy005
> 👋 [@andersy005](https://github.com/andersy005) [@zarzycki](https://github.com/zarzycki) - Would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

my plate's pretty full at the moment, so i need to sit this one out
