### 2025-10-01 #9130 @editorialbot
[REVIEW]: visualCaseGen: An SMT-based Experiment Configurator for Community Earth System Model
**Submitting author:** @alperaltuntas (<a href="http://orcid.org/0000-0003-1708-9518">Alper Altuntas</a>)
**Repository:** https://github.com/ESMCI/visualCaseGen
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.3
