+++ tests/1_unit/test_cores_case_creator.py
@@ -21,7 +21,7 @@ def test_calc_cores_based_on_grid_cases():
 
     assert CaseCreator._calc_cores_based_on_grid(300*32) == 128
 
-    assert CaseCreator._calc_cores_based_on_grid(740 * 780) == 768
+    assert CaseCreator._calc_cores_based_on_grid(740 * 780) == 2048
 
 
     
