### 2025-10-10 db3030e
Add min cores thing
