### 2025-10-10 6ae49ac
NTasks & Timestep Changes suggested by Mike (#18)
* Changes requested by Mike

* Bleh

* Test
'

* Update m6c commit

* Add min cores thing

* m6b update
