+++ visualCaseGen/custom_widget_types/case_creator.py
@@ -45,29 +45,7 @@ class CaseCreator:
         """This function is called when the case creation fails. It reverts the changes made
         to the ccs_config xml files."""
 
-        mg = "ccs_config/modelgrid_aliases_nuopc.xml"
-        if (Path(self._cime.srcroot) / f"{mg}.orig").exists():
-            shutil.move(
-                Path(self._cime.srcroot) / f"{mg}.orig",
-                Path(self._cime.srcroot) / f"{mg}"
-            )
-        cg = "ccs_config/component_grids_nuopc.xml"
-        if (Path(self._cime.srcroot) / f"{cg}.orig").exists():
-            shutil.move(
-                Path(self._cime.srcroot) / f"{cg}.orig",
-                Path(self._cime.srcroot) / f"{cg}"
-            )
-
-    def _remove_orig_xml_files(self):
-        """This function is called when the case creation and modification process is successful.
-        It removes the backup xml files created before modifying the ccs_config xml files."""
-
-        mg = "ccs_config/modelgrid_aliases_nuopc.xml"
-        if (Path(self._cime.srcroot) / f"{mg}.orig").exists():
-            os.remove(Path(self._cime.srcroot) / f"{mg}.orig")
-        cg = "ccs_config/component_grids_nuopc.xml"
-        if (Path(self._cime.srcroot) / f"{cg}.orig").exists():
-            os.remove(Path(self._cime.srcroot) / f"{cg}.orig")
+        return # Currently no action is needed 
 
     def _is_non_local(self):
         """Check if the case is being created on a machine different from the one
@@ -138,11 +116,11 @@ class CaseCreator:
         with self._out:
             print(f"{COMMENT}Creating case...{RESET}\n")
 
-        # First, update ccs_config xml files to add custom grid information if needed:
-        self._update_ccs_config(do_exec)
-
         # Run create_newcase
         self._run_create_newcase(caseroot, compset, resolution, do_exec)
+
+        # Add custom grid information to the case if needed:
+        self._update_grids(do_exec)
 
         # Navigate to the case directory:
         with self._out:
@@ -171,11 +149,10 @@ class CaseCreator:
                     f"another case, restart the notebook.{RESET}\n"
                 )
 
-    def _update_ccs_config(self, do_exec):
+    def _update_grids(self, do_exec):
         """Update the modelgrid_aliases and component_grids xml files with custom grid
-        information if needed. This function is called before running create_newcase."""
+        information if needed. This function is called after running create_newcase."""
 
-        # If Custom grid is selected, update modelgrid_aliases and component_grids xml files:
         if cvars["GRID_MODE"].value == "Standard":
             return
         else:
@@ -201,109 +178,8 @@ class CaseCreator:
         if ocn_grid is None:
             raise RuntimeError("No ocean grid specified.")
 
-        self._update_modelgrid_aliases(custom_grid_path, ocn_grid, do_exec)
         self._update_component_grids(custom_grid_path, ocn_grid, ocn_grid_mode, do_exec)
 
-    def _update_modelgrid_aliases(self, custom_grid_path, ocn_grid, do_exec):
-        """Update the modelgrid_aliases xml file with custom resolution information.
-        This function is called before running create_newcase.
-
-        Parameters
-        ----------
-        custom_grid_path : Path
-            The path to the custom grid directory.
-        ocn_grid : str
-            The name of the custom ocean grid.
-        do_exec : bool
-            If True, execute the commands. If False, only print them.
-            """
-
-        resolution_name = custom_grid_path.name
-
-        # Component grid names:
-        atm_grid = cvars["CUSTOM_ATM_GRID"].value
-        lnd_grid = cvars["CUSTOM_LND_GRID"].value
[…]
