### 2025-10-24 ESMCI/visualCaseGen#22 @alperaltuntas
Fix error handling for standard ocn grids
When generating cases with custom component sets, the program occasionally throws the following error when generating the command for case building:

```
int() argument must be a string, a bytes-like object or a real number, not 'NoneType'
```

Reported by @fzhao70

https://github.com/openjournals/joss-reviews/issues/9130
