### 2025-10-24 69b14f3
add libxml2 dependency
