+++ visualCaseGen/custom_widget_types/case_creator.py
@@ -235,9 +235,9 @@ class CaseCreator:
 
             xmlchange("LND_GRID", cvars["CUSTOM_LND_GRID"].value, do_exec, self._is_non_local(), self._out)
 
-            xmlchange("ATM_DOMAIN_MESH", cvars["CUSTOM_ATM_DOMAIN_MESH"].value, do_exec, self._is_non_local(), self._out)
+            xmlchange("ATM_DOMAIN_MESH", os.path.expandvars(self._cime.domains["atm"][cvars["CUSTOM_ATM_GRID"].value].mesh), do_exec, self._is_non_local(), self._out)
 
-            xmlchange("LND_DOMAIN_MESH", cvars["CUSTOM_LND_DOMAIN_MESH"].value, do_exec, self._is_non_local(), self._out)
+            xmlchange("LND_DOMAIN_MESH",  os.path.expandvars(self._cime.domains["lnd"][cvars["CUSTOM_LND_GRID"].value].mesh), do_exec, self._is_non_local(), self._out)
 
 
         lnd_grid_mode = cvars["LND_GRID_MODE"].value
