### 2025-11-13 2ef79c2
Fix case generation for standard MOM6 grids. (#23)
* add a test that catches the issue with std mom6 grid:

Case creator attemps to set NTASKS_OCN for by checking
OCN_NX and OCN_NY but those variables are not set for
standard MOM6 grids.

* fix the issue with standard MOM6 grids:

Case creator attemps to set NTASKS_OCN for by checking
OCN_NX and OCN_NY but those variables are not set for
standard MOM6 grids.

While we can retrieve them for
standard grids too, we should still not change NTASKS
for standard ocn grids, because they have associated,
and optimized NTASKS values set in config_pes fi
[…]
