### 2025-11-17 79280c5
Add custom runoff grid
This commit adds the ability to specify a custom runoff grid.
This includes changes to the grid options, relational constraints, and
the grid stages. The custom runoff grid stage is added to the stage
pipeline, right after the custom lnd grid stage.
