+++ visualCaseGen/specs/relational_constraints.py
@@ -24,6 +24,7 @@ def get_relational_constraints(cvars):
     OCN_NX = cvars['OCN_NX']; OCN_NY = cvars['OCN_NY']; OCN_LENX = cvars['OCN_LENX']; OCN_LENY = cvars['OCN_LENY']
     LND_GRID_MODE = cvars['LND_GRID_MODE']; LND_SOIL_COLOR = cvars['LND_SOIL_COLOR']; LND_DOM_PFT = cvars['LND_DOM_PFT']
     LND_MAX_SAT_AREA = cvars['LND_MAX_SAT_AREA']; LND_STD_ELEV = cvars['LND_STD_ELEV']
+    CUSTOM_ROF_GRID = cvars['CUSTOM_ROF_GRID']
 
     # Return a dictionary of constraints where keys are the z3 boolean expressions corresponding to the constraints
     # and values are error messages to be displayed when the constraint is violated.
@@ -165,7 +166,14 @@ def get_relational_constraints(cvars):
             "Max fraction of saturated area must be set to a value between 0 and 1.",
 
         LND_STD_ELEV >= 0.0:
-            "Standard deviation of elevation must be a nonnegative number."
+            "Standard deviation of elevation must be a nonnegative number.",
+
+        # Custom rof grid constraints ------------------
+        Implies(In(CUSTOM_ROF_GRID, ["JRA025", "rx1"]), COMP_ROF=="drof"):
+            "JRA025 and rx1 runoff grids can only be selected if DROF is the runoff component.",
+
+        Contains(COMP_ROF_OPTION, "JRA") == (CUSTOM_ROF_GRID == "JRA025"):
+            "JRA data runoff forcing can only be used iff the custom runoff grid is set to JRA025.",
 
         #### Assertions to stress-test the CSP solver
 
