### 2025-11-23 eee8995
improve _process_domain_constraints and update relational constraints
Previously, _process_domain_constraints generated the disjunction
of compset and not_compset constraints read in from ccs_config
xml files, but we really want the conjunction of not_compset constraints
alongside the disjjunction of compset constraints. This commit
achieves that. Also made a minor change in custom runoff grid relational
specification.
