### 2025-12-04 e32a9b2
introduce retrieve_maps and get_mesh_path mehods in CIME_interface
