### 2025-12-04 6be5ccf
Introduce runoff to ocean mapping stage
CHanges include:
 - New runoff to ocean mapping flag and parameters.
 - Runoff mapping generator widget
 - Relational constraints
 - New guarded mapping stage following runoff stage (only activates
   if mapping is necessary.
 - If a standard mapping file is found between the selected runoff
   and ocean grids, the user has the option to use that standard map.
 - xml changes for mapping files.
