### 2025-12-04 c44cf05
relax not_compset constraint for default grids
