### 2025-12-08 114b537
Remove user nl and xml query
