+++ _sources/grid.rst.txt
@@ -1,0 +1,188 @@
+Stage 2/3: Grid
+===============
+
+The second major step in configuring your CESM case is choosing a resolution, which
+is a specific set of grids for each active and data model in the compset. You may 
+select either a standard, out-of-the-box resolution or create a custom one by combining
+existing model grids. In `Custom` mode, you can also generate custom model grids for CLM and/or MOM6
+using auxiliary tools included in visualCaseGen. Begin by selecting between `Standard`
+and `Custom` grid modes.
+
+.. image:: assets/Stage2_1.png
+
+.. note:: In CESM terminology, *resolution* and *grid* are often used interchangeably,
+   both referring to the combination of model grids used in a CESM simulation. Unless 
+   specifically noted as a *model grid* (i.e., a grid unique to a particular component,
+   such as the ocean grid), the term *grid* in this context should be understood as
+   *resolution*, meaning the full collection of *model grids* used in a particular CESM case.
+
+Standard Grids
+------------------
+
+Select from the available list of resolutions (combinations of model grids) below.
+Resolutions known to be incompatible with your chosen compset have been omitted 
+from this list. Use the search box to refine the list further. For exact matches,
+use double quotes; otherwise, the search will display all grids containing one 
+or more of the search terms.
+
+.. image:: assets/Stage2_2.png
+
+After selecting a grid, visualCaseGen will advance to the `Launch` stage, where
+you can create your CESM case using the chosen compset and grid configuration.
+
+Custom Grids
+------------------
+
+In Custom Grid mode, you can build a custom grid by mixing and matching standard 
+model grids or generating new MOM6 and/or CLM grids with specialized tools that come with visualCaseGen.
+Start by specifying a path to save the new grid files and a name to refer to this
+new grid in the configuration process and beyond.
+
+.. image:: assets/Stage2_3.png
+
+After clicking `Select`, a file browser will open to help you locate your preferred
+directory for saving the new grid files. Once the directory is selected, enter the
+new grid name in the text box at the top right and click `Select` to proceed.
+
+.. image:: assets/Stage2_4.png
+
+Atmosphere Grid
+~~~~~~~~~~~~~~~
+
+Next, choose an atmosphere grid from the list of compatible options based on the
+compset you selected in the `Compset` stage. You may use the search box to filter
+the list further. Note that, currently, visualCaseGen does not support the creation
+of custom atmosphere grids. Therefore, you can only select from the standard atmosphere
+grids available in the list. The selected standard atmosphere grid will then be included
+in the custom resolution that you are building.
+
+.. image:: assets/Stage2_5.png
+
+Ocean Grid
+~~~~~~~~~~
+
+For the ocean grid, if MOM6 is selected as the ocean model, you can either select a standard
+ocean grid or create a new MOM6 grid. When creating a new MOM6 grid, you'll specify parameters
+such as grid extent and resolution, after which you'll be directed to a separate notebook that
+uses the `mom6_bathy` tool to generate the new grid and bathymetry.
+
+If using a standard ocean grid, select one from the list compatible with your chosen compset
+and atmosphere grid. If creating a new MOM6 grid, complete the required parameters, then proceed
+to launch the `mom6_bathy` tool for final customization.
+
+.. image:: assets/Stage2_6.png
+
+After specifying all ocean grid parameters, click `Launch mom6_bathy`. This will open an 
+auto-generated Jupyter notebook where you can fine-tune the ocean grid, topography, and vertical grid.
+You can then generate the corresponding MOM6 input files, which will be saved under `ocnice` subdirectory
+within the directory you specified earlier for saving the new grid files.
+For more details on mom6_bathy, refer to its documentation: https://ncar.github.io/mom6_bathy/
+
[…]
