+++ visualCaseGen/custom_widget_types/case_creator.py
@@ -23,7 +23,7 @@ BPOINT = "\u2022"
 class CaseCreator:
     """The base class for CaseCreatorWidget. Here, backend functionalities are implemented."""
 
-    def __init__(self, cime, output=None, allow_xml_override=False):
+    def __init__(self, cime, output=None, allow_xml_override=False, add_grids_to_ccs_config = True):
         """Initialize CaseCreator object.
 
         Parameters
@@ -40,15 +40,13 @@ class CaseCreator:
         self._cime = cime
         self._out = DummyOutput() if output is None else output
         self._allow_xml_override = allow_xml_override
-        self._assign_grids_through_ccs_config = is_ccs_config_writeable(cime) # By default, visualCaseGen assigns grids through ccs_config, if not possible (which can happen if the user does not own the sandbox), it applies the change through xml case changes.
-        if not self._assign_grids_through_ccs_config:
-            with self._out:
-                print(f"{COMMENT}ccs_config is not writeable, so case grids will be assigned through xml changes{RESET}\n")
+        self._add_grids_to_ccs_config = add_grids_to_ccs_config # By default, visualCaseGen assigns grids through ccs_config, if not possible (which can happen if the user does not own the sandbox), it is possible to apply grid changes through xmlchanges instead.
+        assert is_ccs_config_writeable or not add_grids_to_ccs_config, "Cannot write to ccs_config xml files. Please set add_grids_to_ccs_config to False to apply grid changes through xmlchanges."
 
     def revert_launch(self, do_exec=True):
         """This function is called when the case creation fails. It reverts the changes made
         to the ccs_config xml files."""
-        if self._assign_grids_through_ccs_config:
+        if self._add_grids_to_ccs_config:
             mg = "ccs_config/modelgrid_aliases_nuopc.xml"
             if (Path(self._cime.srcroot) / f"{mg}.orig").exists():
                 shutil.move(
@@ -134,10 +132,10 @@ class CaseCreator:
         if cvars["GRID_MODE"].value == "Standard":
             resolution = cvars["GRID"].value
         elif cvars["GRID_MODE"].value == "Custom":
-            if self._assign_grids_through_ccs_config:
+            if self._add_grids_to_ccs_config:
                 resolution = Path(cvars["CUSTOM_GRID_PATH"].value).name
             else:
-                resolution = "inputdir" # Set to a default visualCaseGen Resolution since grids are changed through xml changes
+                resolution = "USER_RES" # Set to a default visualCaseGen Resolution since grids are changed through xml changes
         else:
             raise RuntimeError(f"Unknown grid mode: {cvars['GRID_MODE'].value}")
 
@@ -146,15 +144,15 @@ class CaseCreator:
             print(f"{COMMENT}Creating case...{RESET}\n")
 
         # First, update ccs_config xml files to add custom grid information if needed:
-        if self._assign_grids_through_ccs_config:
+        if self._add_grids_to_ccs_config:
             self._update_ccs_config(do_exec)
 
         # Run create_newcase
         self._run_create_newcase(caseroot, compset, resolution, do_exec)
 
         # If we don't pick the grids through ccs_config, use xml changes
-        if not self._assign_grids_through_ccs_config: 
-            self._update_grids(do_exec)
+        if not self._add_grids_to_ccs_config: 
+            self._update_grids_via_xmlchange(do_exec)
 
[…]
