+++ _sources/ridge.rst.txt
@@ -1,0 +1,315 @@
+Ridge World
+==============================
+
+This example provides step-by-step guidance on generating a coupled idealized configuration
+consisting of an aquaplanet with polar land caps and a narrow land ridge extending between
+the two poles. This setup is similar to the configuration used in `Wu et al (2021) <https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2021GL093966>`_ .
+The example demonstrates how to use the visualCaseGen GUI to select CESM components, configure
+custom ocean and land input files, and set up the final case.
+
+Unlike standard CESM cases, this example involves significant modifications, including
+customizations to the ocean grid, bathymetry, and land surface
+properties.
+
+
+.. note:: A video demonstration of this example is available `at this link <https://www.youtube.com/watch?v=YwBqcIlW_fg>`_.
+
+Stage 0: Open visualCaseGen
+---------------------------
+
+Follow the instructions in the :ref:`Open` to open visualCaseGen in your Jupyter notebook environment.
+
+Stage 1: Select Compset
+-----------------------
+
+After having executed the cell with the command `from visualCaseGen import gui; gui`,
+and clicking the **Start** button, you will see the main interface of visualCaseGen. 
+
+Click the **Custom** button to proceed with creating a custom compset for our idealized
+Ridge World configuration.
+
+.. image:: assets/stage1_1.png
+
+Initialization time
+~~~~~~~~~~~~~~~~~~~
+
+Once the `Custom` button is clicked, you will be prompted to select the initialization time.
+For this example, select `1850` as the initialization time.
+
+.. image:: assets/stage1_5.png
+
+
+Models
+~~~~~~
+
+Once you've selected the initialization time, visualCaseGen will prompt you to select the
+models for each component class. For this ridge world case, select the following component options:
+`cam` as the atmosphere;
+`clm` as the land component; `cice` as the ice component; `mom` as the ocean component, `srof` 
+(i.e., stub run off) as the river component; `sglc` (i.e., stub land ice) as the land ice component;
+and, `swav` (i.e. stub wave) as the wave component. After all the selections are made, the 
+model matrix should look like:
+
+.. image:: assets/ridge1.png
+
+Model Physics
+~~~~~~~~~~~~~
+
+Having selected the models, you will proceed to select the physics options for each model. The physics
+settings determine the complexity of each model component and impact computational requirements.
+For this example, select the following physics options which are based on the selections
+made in `Wu et al (2021) <https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2021GL093966>`_ :
+
+.. image:: assets/ridge2.png
+
+Component Options (Modifiers)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+You are now ready to finalize the compset by selecting optional physics modifiers. Navigate between
+tabs for each component options to select the desired options. For this example, select `(none)` for all the
+components except for the `clm` component, which should be set to the satellite phenology (SP) mode.
+This options enables various aspects of the land model such as
+leaf-area index are prescribed as opposed to being prognosed interactively by the land biogeochemistry.
+
+.. image:: assets/ridge3.png
+
+Stage 2: Grid
+----------------------
+Having completed the compset configuration, you will now proceed to the `Grid` stage. This can
+either be a "Standard" grid, i.e., a ready-made grid that comes with CESM, or a "Custom" grid
+where you can mix-and-match existing model grids or create new ones.
+
+.. image:: assets/Stage2_1.png
+
[…]
