### 2026-02-20 ESMCI/visualCaseGen#31 @mengqi-z
JOSS Paper - Proofreading
Hi @alperaltuntas,

This is for JOSS paper review openjournals/joss-reviews#9130

Please address the following points in `paper.md`:

1. Use inline code formatting for all software and tool names in `paper.md`. For example, use \`visualCaseGen\`.
2. For the code snippet after line 51, I suggest using a code block (e.g., \```code block\```) rather than a screenshot. The JOSS paper compiler will render the code block appropriately in code format.

Thank you.
