### 2026-02-20 8a33814
use inline code formatting for software and tool names
