### 2026-02-20 d9eb85d
use code snipped for constraints
