### 2026-02-20 ESMCI/visualCaseGen#31 @alperaltuntas
Thank you @mengqi-z! I have just pushed the following two commits to address these issues:

8a3381460dc1bd7a18c986cac3ae129bc0724357
d9eb85d334e3595828a3343ca4932f10094e1374
