### 2026-02-20 #9130 @mengqi-z
Hi @alperaltuntas - Could you please complete the following items:

- [ ] Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
- [ ] Archive the release on Zenodo/figshare/etc and **post the DOI here**.

Then I can move forward with accepting the paper. Thanks!
