### 2026-02-20 #9130 @brian-rose
Dear @mengqi-z, @alperaltuntas and author team,

I'd to apologize for holding this review up for so long. I found it extraordinarily difficult to carve out enough time to get the work done. I should have requested to be removed from the reviewer list earlier.

I'm glad that the paper can move forward, and I hope to review for JOSE again in the future.
