### 2026-02-23 #9130 @mengqi-z
@brian-rose - Thank you for your response. I understand how busy schedules can get, though your expertise would have been valuable to this review. I look forward to your future contributions to JOSS.
