### 2026-03-02 #9130 @alperaltuntas
Hi @mengqi-z

 > Double check authors and affiliations (including ORCIDs).

 Done.

> Make a release of the software with the latest changes

Done: v1.0.0

> Archive the release on Zenodo/figshare/etc and post the DOI here.

Done: [10.5281/zenodo.18841887](https://doi.org/10.5281/zenodo.18841887)

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done.

> Make sure that the license listed for the archive is the same as the software license.

Done.

Thanks!
