### 2026-03-03 #9130 @mengqi-z
@alperaltuntas - Thanks!

Could you please update the **title and author list** in your metadata for the Zenodo archive? They need to match the JOSS paper’s title and authors exactly.

Just a note that the tagged release version is `main_1.0.0` from what I see [here](https://github.com/ESMCI/visualCaseGen/tree/main_1.0.0), which will be used instead of `v1.0.0`.
