### 2026-03-04 #9130 @mengqi-z
@alperaltuntas – I’ve recommended your submission for acceptance. Thanks so much for submitting to JOSS! From here, our Editor-in-Chief will step in to do some final checks before the paper can be officially accepted and published. Congrats and great job! 🎉

@lars2015, @fzhao70 – Huge thanks to all of you for reviewing this paper! It’s been great working with you during this process. I look forward to seeing your reviews for JOSS in the future!
